+++ src/VM.h
@@ -273,10 +273,10 @@ private:
     }
 
     void initializeDynamicCode();
-    Object evaluateSafe(Object* code, int codeSize, bool isCompiler = false);
+    Object evaluateSafe(Object* code, size_t codeSize, bool isCompiler = false);
     Object evaluateSafe(Code* code);
     Object evaluateSafe(Vector* code);
-    Object evaluateUnsafe(Object* code, int codeSize, bool isCompiler = false);
+    Object evaluateUnsafe(Object* code, size_t codeSize, bool isCompiler = false);
     Object evaluateUnsafe(Vector* code, bool isCompiler = false);
 
     void makeCallFrame(Object* pc)
@@ -287,7 +287,7 @@ private:
         push(Object::makeObjectPointer(fp_));
     }
 
-    Object* getDirectThreadedCode(/*const*/ Object* code, int length, bool isCompiler = false);
+    Object* getDirectThreadedCode(/*const*/ Object* code, size_t length, bool isCompiler = false);
     Object runLoop(Object* code, jmp_buf returnPoint, bool returnTable = false);
 
     Object ac_;  // accumulator     register
